Compare shiftable "ROMs" instead of generating a fake ELF
Previously, the build process generated two products:
- a "fake" ELF, which contained the final matching binary
- the matching "ELF", which we `objcopy`'d out of the "fake" ELF
The result was a completely matching ELF, but not a legitimate one.
The compiled code was simply injected into the original ELF metadata.

This was perfectly viable for our purposes, but would cause problems
later if someone wanted to actually modify the code (it wasn't shiftable).

With this commit, the final products are now:
- a *nonmatching* but legitimate ELF
  - This ELF contains matching code/data, but the ELF metadata does not
    match the original binary. It can still be substituted in place of the
    original binary.
- a "ROM" which only contains the code/data sections
  - We compare this to a similarly generated "ROM" from the original binary.
    If this is matching, our ELF is _effectively_ matching even though the
    metadata does not.

In summary, we no longer try to match the original ELF's debug symbols or
metadata. This wasn't really our goal anyway; we only care about matching
code/data/functionality.